Dropped `takes_args` and introduced `optional_arg`
We dropped `takes_args` from `class Option`.

Now we rely on the `complete` parameter to find out if an option takes
arguments.

If `complete` is not `None`, the option takes an argument.

If an option takes an argument, but has no appropriate completion, then
`["none"]` should be used.

If an option takes an optional argument, then `optional_arg=True` should
be used.